code-voyager/pyproject.toml
[project]
name = "voyager"
version = "0.2.0"
description = "AI assistant memory system that works with any CLI and IDE"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "claude-agent-sdk>=0.1.18",
  "jsonschema>=4.25.1",
  "pyyaml>=6.0",
  "typer>=0.15.0",
  "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
# ColBERT skill retrieval
colbert = ["ragatouille>=0.0.8"]
# Backward compatibility alias
retrieval = ["ragatouille>=0.0.8"]
# OpenAI AI provider
openai = ["openai>=1.0.0"]
# Ollama AI provider
ollama = ["httpx>=0.25.0"]
# OpenRouter AI provider
openrouter = ["httpx>=0.25.0"]
# OpenAI-compatible providers
openai-compatible = ["openai>=1.0.0"]
# LSP server
lsp = ["pygls>=1.1.0", "lsprotocol>=2023.0.0"]
# Google Gemini AI provider
gemini = ["google-generativeai>=0.3.0"]
# Cohere AI provider
cohere = ["cohere>=4.0.0"]
# SimpleMem integration
simplemem = ["lancedb>=0.3.0", "dateparser>=1.1.0"]
# All AI providers
all-providers = ["openai>=1.0.0", "httpx>=0.25.0", "google-generativeai>=0.3.0", "cohere>=4.0.0"]
# Full installation with all features
full = [
  "ragatouille>=0.0.8",
  "openai>=1.0.0",
  "httpx>=0.25.0",
  "pygls>=1.1.0",
  "lsprotocol>=2023.0.0",
  "google-generativeai>=0.3.0",
  "cohere>=4.0.0",
  "lancedb>=0.3.0",
  "dateparser>=1.1.0",
]
# Development dependencies
dev = ["ruff>=0.8.0", "pytest>=8.0.0", "pytest-asyncio>=0.21.0"]

[project.scripts]
voyager = "voyager.cli:main"
voyager-lsp = "voyager.lsp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voyager"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
